refactor(app): dependency injection with hilt added. NetworkModule.kt added.
