+++ app/build.gradle
@@ -4,7 +4,7 @@ plugins {
     id("kotlin-android-extensions")
     id("kotlin-kapt")
     id("androidx.navigation.safeargs")
-
+    id("dagger.hilt.android.plugin")
 }
 
 android {
@@ -45,8 +45,17 @@ dependencies {
     implementation(Dependencies.material)
     implementation(Dependencies.timber)
     implementation(Dependencies.constraintLayout)
+
+    implementation(Dependencies.hilt)
+    kapt(Dependencies.hiltKapt)
+
+
     testImplementation(Dependencies.jUnit)
 
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+}
+
+kapt {
+    correctErrorTypes true
 }
