+++ app/src/main/java/com/github/mvibase/App.kt
@@ -1,0 +1,13 @@
+package com.github.mvibase
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application() {
+
+    override fun onCreate() {
+        super.onCreate()
+    }
+
+}
