+++ app/src/main/java/com/github/mvibase/di/NetworkModule.kt
@@ -1,0 +1,12 @@
+package com.github.mvibase.di
+
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.android.components.ActivityComponent
+import dagger.hilt.android.internal.managers.ApplicationComponentManager
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+class NetworkModule {
+}
