+++ build.gradle
@@ -19,11 +19,11 @@ buildscript {
 
         classpath(BuildPlugins.androidGradle)
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$Versions.kotlin"
-
-        classpath("com.jfrog.bintray.gradle:gradle-bintray-plugin:1.7.3")
+        classpath "com.google.dagger:hilt-android-gradle-plugin:$Versions.hilt"
+        classpath("com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4")
         classpath("android.arch.navigation:navigation-safe-args-gradle-plugin:1.0.0")
         classpath("com.google.gms:google-services:4.3.10")
-        classpath("com.novoda:bintray-release:0.8.1")
+        classpath("com.novoda:bintray-release:0.9.1")
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
